feat(transformer_lens): add ref cache
